fix: LottoNumbers의 getter 제거
