+++ src/main/java/lotto/model/LottoWinner.java
@@ -6,39 +6,23 @@ import java.util.List;
 
 public class LottoWinner extends LottoNumbers {
 
-    private LottoNumber bonusNumber;
+    private final LottoNumber bonusNumber;
 
     public LottoWinner(List<LottoNumber> lottoNumbers, LottoNumber bonusNumber) {
         super(lottoNumbers);
         this.bonusNumber = bonusNumber;
     }
-
-    public LottoWinner(List<LottoNumber> lottoNumbers) {
-        super(lottoNumbers);
-    }
-
-    public LottoRank match(LottoNumbers lottoNumbers) {
+    
+    public LottoRank match(LottoNumbers matchNumbers) {
         Number count = new Number();
-        for (LottoNumber lottoNumber : lottoNumbers.getLottoNumbers()) {
-            match(lottoNumber, count);
+
+        for (LottoNumber lottoNumber : super.lottoNumbers) {
+            matchNumbers.match(lottoNumber, count);
         }
 
-        boolean isBonus = bonusMatch(lottoNumbers);
+        boolean isBonus = matchNumbers.contains(bonusNumber);
+
         return LottoRank.valueOf(count, isBonus);
-    }
-
-    public void match(LottoNumber lottoNumber, Number count) {
-        if (contains(lottoNumber)) {
-            count.plus();
-        }
-    }
-
-    public boolean bonusMatch(LottoNumbers lottoNumbers) {
-        return lottoNumbers.getLottoNumbers().contains(bonusNumber);
-    }
-
-    public boolean contains(LottoNumber lottoNumber) {
-        return this.lottoNumbers.contains(lottoNumber);
     }
 
 }
