+++ src/test/java/common/model/NumberTest.java
@@ -12,13 +12,6 @@ public class NumberTest {
 
 
     @Test
-    @DisplayName("생성자 테스트")
-    void constructor() {
-        Number number = new Number("1");
-        assertThat(number.getNumber()).isEqualTo(1);
-    }
-
-    @Test
     @DisplayName("equals 테스트")
     void equals() {
         Number number = new Number("1");
