+++ src/test/java/lotto/model/LottoNumbersTest.java
@@ -1,6 +1,5 @@
 package lotto.model;
 
-import lotto.application.Constant;
 import lotto.factory.LottoNumbersFactory;
 import lotto.generator.LottoNumberGenerator;
 import org.junit.jupiter.api.DisplayName;
@@ -14,7 +13,7 @@ class LottoNumbersTest {
     @DisplayName("로또 번호 자동생성")
     void autoCreate() {
         LottoNumbers lottoNumbers = LottoNumbersFactory.autoCreateNumbers(new LottoNumberGenerator());
-        assertThat(lottoNumbers.getLottoNumbers()).hasSize(Constant.LOTTO_NUMBERS_SIZE);
+        assertThat(lottoNumbers).isInstanceOf(LottoNumbers.class);
     }
 
 }
