+++ src/test/java/lotto/model/LottoRankTest.java
@@ -4,19 +4,11 @@ import common.model.Number;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.ValueSource;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 class LottoRankTest {
-
-    @ParameterizedTest
-    @CsvSource(value = {"6,2000000000", "5,30000000", "4,50000", "3,5000"})
-    @DisplayName("로또 당첨금 확인 테스트")
-    void lottoAmount(int matchCount, int amount) {
-        assertThat(LottoRank.valueOf(new Number(matchCount)).getAmount()).isEqualTo(amount);
-    }
 
     @Test
     @DisplayName("2등 테스트")
