+++ src/test/java/lotto/model/LottoReportTest.java
@@ -1,6 +1,7 @@
 package lotto.model;
 
 import lotto.factory.LottoFactory;
+import lotto.factory.LottoNumberFactory;
 import lotto.factory.LottoNumbersFactory;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -14,7 +15,8 @@ class LottoReportTest {
     void calculateRateOfRevenue() {
         LottoNumbers lottoNumbers = LottoNumbersFactory.manualCreateNumbers("1, 2, 3, 4, 5, 6");
         Lotto lotto = LottoFactory.manualCreateSingleLotto(lottoNumbers);
-        LottoWinner winner = new LottoWinner(LottoNumbersFactory.manualCreateNumberList("1, 2, 3, 11, 12, 13"));
+        LottoNumber bonus = LottoNumberFactory.manualCreateNumber(7);
+        LottoWinner winner = new LottoWinner(LottoNumbersFactory.manualCreateNumberList("1, 2, 3, 11, 12, 13"), bonus);
 
         LottoReport report = new LottoReport(lotto, winner);
